Fix maybe-merge-trees
