+++ lib/engine-addon.js
@@ -13,6 +13,7 @@ var defaultsDeep = require('lodash/defaultsDeep');
 var calculateCacheKeyForTree = require('calculate-cache-key-for-tree');
 var Addon = require('ember-cli/lib/models/addon');
 var memoize = require('./utils/memoize');
+var maybeMergeTrees = require('./utils/maybe-merge-trees');
 var p = require('ember-cli-preprocess-registry/preprocessors');
 var CompactReexports = require('babel-plugin-compact-reexports');
 var shouldCompactReexports = require('./utils/should-compact-reexports');
@@ -285,22 +286,6 @@ var buildCompleteJSTree = memoize(function buildCompleteJSTree() {
 
   return mergeTrees([vendorJSTree, engineAppTree], { overwrite: true });
 });
-
-/*
-  Small helper function to reduce extra effort when invoking
-  `mergeTrees([])` or `mergeTrees([something])`.
-
-  @param {Array} _inputTrees an array of potential trees to merge
-  @return {Tree|undefined}
-*/
-function maybeMergeTrees(_inputTrees, options) {
-  var inputTrees = _inputTrees.filter(Boolean);
-  if (inputTrees.length > 1) {
-    return mergeTrees(inputTrees, options);
-  } else {
-    return inputTrees[0];
-  }
-}
 
 var buildEngineStyleTree = memoize(function buildEngineStyleTree() {
   // gather engines own `addon/styles` and its dependencies `app/styles`
